[AIRFLOW-5950] AIP-21 Change import paths for "apache/cassandra" modules
