+++ airflow/contrib/sensors/cassandra_record_sensor.py
@@ -16,43 +16,16 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from airflow.contrib.hooks.cassandra_hook import CassandraHook
-from airflow.sensors.base_sensor_operator import BaseSensorOperator
-from airflow.utils.decorators import apply_defaults
 
+"""This module is deprecated. Please use `airflow.providers.apache.cassandra.sensors.record`."""
 
-class CassandraRecordSensor(BaseSensorOperator):
-    """
-    Checks for the existence of a record in a Cassandra cluster.
+import warnings
 
-    For example, if you want to wait for a record that has values 'v1' and 'v2' for each
-    primary keys 'p1' and 'p2' to be populated in keyspace 'k' and table 't',
-    instantiate it as follows:
+# pylint: disable=unused-import
+from airflow.providers.apache.cassandra.sensors.record import CassandraRecordSensor  # noqa
 
-    >>> cassandra_sensor = CassandraRecordSensor(table="k.t",
-    ...                                          keys={"p1": "v1", "p2": "v2"},
-    ...                                          cassandra_conn_id="cassandra_default",
-    ...                                          task_id="cassandra_sensor")
-
-    :param table: Target Cassandra table.
-        Use dot notation to target a specific keyspace.
-    :type table: str
-    :param keys: The keys and their values to be monitored
-    :type keys: dict
-    :param cassandra_conn_id: The connection ID to use
-        when connecting to Cassandra cluster
-    :type cassandra_conn_id: str
-    """
-    template_fields = ('table', 'keys')
-
-    @apply_defaults
-    def __init__(self, table, keys, cassandra_conn_id, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.cassandra_conn_id = cassandra_conn_id
-        self.table = table
-        self.keys = keys
-
-    def poke(self, context):
-        self.log.info('Sensor check existence of record: %s', self.keys)
-        hook = CassandraHook(self.cassandra_conn_id)
-        return hook.record_exists(self.table, self.keys)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.apache.cassandra.sensors.record`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
