+++ docs/conf.py
@@ -225,6 +225,8 @@ exclude_patterns = [
     '_api/airflow/providers/amazon/index.rst',
     '_api/airflow/providers/amazon/aws/index.rst',
     '_api/airflow/providers/amazon/aws/example_dags',
+    '_api/airflow/providers/apache/index.rst',
+    '_api/airflow/providers/apache/cassandra/index.rst',
     'autoapi_templates',
     'howto/operator/gcp/_partials',
 ]
