+++ scripts/ci/pylint_todo.txt
@@ -8,7 +8,6 @@
 ./airflow/contrib/hooks/azure_cosmos_hook.py
 ./airflow/contrib/hooks/azure_data_lake_hook.py
 ./airflow/contrib/hooks/azure_fileshare_hook.py
-./airflow/contrib/hooks/cassandra_hook.py
 ./airflow/contrib/hooks/datadog_hook.py
 ./airflow/contrib/hooks/dingding_hook.py
 ./airflow/contrib/hooks/discord_webhook_hook.py
@@ -45,7 +44,6 @@
 ./airflow/contrib/operators/emr_terminate_job_flow_operator.py
 ./airflow/contrib/operators/file_to_wasb.py
 ./airflow/contrib/operators/grpc_operator.py
-./airflow/contrib/operators/hipchat_operator.py
 ./airflow/contrib/operators/hive_to_dynamodb.py
 ./airflow/contrib/operators/jenkins_job_trigger_operator.py
 ./airflow/contrib/operators/jira_operator.py
@@ -72,7 +70,6 @@
 ./airflow/contrib/operators/sftp_to_s3_operator.py
 ./airflow/contrib/operators/slack_webhook_operator.py
 ./airflow/contrib/operators/snowflake_operator.py
-./airflow/contrib/operators/sns_publish_operator.py
 ./airflow/contrib/operators/spark_jdbc_operator.py
 ./airflow/contrib/operators/spark_sql_operator.py
 ./airflow/contrib/operators/spark_submit_operator.py
@@ -86,8 +83,6 @@
 ./airflow/contrib/sensors/aws_glue_catalog_partition_sensor.py
 ./airflow/contrib/sensors/azure_cosmos_sensor.py
 ./airflow/contrib/sensors/bash_sensor.py
-./airflow/contrib/sensors/cassandra_record_sensor.py
-./airflow/contrib/sensors/cassandra_table_sensor.py
 ./airflow/contrib/sensors/celery_queue_sensor.py
 ./airflow/contrib/sensors/datadog_sensor.py
 ./airflow/contrib/sensors/emr_base_sensor.py
@@ -192,6 +187,9 @@
 ./airflow/providers/amazon/aws/operators/athena.py
 ./airflow/providers/amazon/aws/sensors/athena.py
 ./airflow/providers/amazon/aws/sensors/redshift.py
+./airflow/providers/apache/cassandra/hooks/cassandra.py
+./airflow/providers/apache/cassandra/sensors/record.py
+./airflow/providers/apache/cassandra/sensors/table.py
 ./airflow/sensors/base_sensor_operator.py
 ./airflow/sensors/external_task_sensor.py
 ./airflow/sensors/hdfs_sensor.py
@@ -264,38 +262,28 @@
 ./airflow/www/validators.py
 ./airflow/www/views.py
 ./airflow/www/widgets.py
+./dags/test_dag.py
 ./scripts/perf/scheduler_ops_metrics.py
 ./tests/contrib/hooks/test_aws_hook.py
 ./tests/contrib/hooks/test_azure_container_instance_hook.py
 ./tests/contrib/hooks/test_azure_cosmos_hook.py
-./tests/contrib/hooks/test_cassandra_hook.py
 ./tests/contrib/hooks/test_databricks_hook.py
 ./tests/contrib/hooks/test_imap_hook.py
 ./tests/contrib/hooks/test_jira_hook.py
 ./tests/contrib/hooks/test_mongo_hook.py
 ./tests/contrib/hooks/test_openfaas_hook.py
-./tests/contrib/hooks/test_opsgenie_alert_hook.py
-./tests/contrib/hooks/test_slack_webhook_hook.py
 ./tests/contrib/hooks/test_snowflake_hook.py
 ./tests/contrib/hooks/test_spark_sql_hook.py
-./tests/contrib/hooks/test_spark_submit_hook.py
-./tests/contrib/hooks/test_sqoop_hook.py
-./tests/contrib/hooks/test_zendesk_hook.py
 ./tests/contrib/operators/test_azure_container_instances_operator.py
 ./tests/contrib/operators/test_azure_cosmos_insertdocument_operator.py
 ./tests/contrib/operators/test_druid_operator.py
 ./tests/contrib/operators/test_ecs_operator.py
 ./tests/contrib/operators/test_emr_add_steps_operator.py
 ./tests/contrib/operators/test_emr_create_job_flow_operator.py
-./tests/contrib/operators/test_oracle_to_azure_data_lake_transfer.py
-./tests/contrib/operators/test_qubole_check_operator.py
-./tests/contrib/operators/test_redis_publish_operator.py
 ./tests/contrib/operators/test_s3_copy_object_operator.py
 ./tests/contrib/operators/test_s3_delete_objects_operator.py
-./tests/contrib/operators/test_s3_to_sftp_operator.py
 ./tests/contrib/operators/test_sftp_operator.py
 ./tests/contrib/operators/test_snowflake_operator.py
-./tests/contrib/operators/test_spark_submit_operator.py
 ./tests/contrib/operators/test_ssh_operator.py
 ./tests/contrib/sensors/test_bash_sensor.py
 ./tests/contrib/sensors/test_emr_step_sensor.py
@@ -304,11 +292,9 @@
 ./tests/contrib/sensors/test_redis_pub_sub_sensor.py
 ./tests/contrib/sensors/test_weekday_sensor.py
 ./tests/core.py
-./tests/executors/test_base_executor.py
 ./tests/executors/test_celery_executor.py
 ./tests/executors/test_dask_executor.py
 ./tests/hooks/test_docker_hook.py
-./tests/hooks/test_druid_hook.py
 ./tests/hooks/test_hive_hook.py
 ./tests/hooks/test_http_hook.py
 ./tests/hooks/test_oracle_hook.py
@@ -326,13 +312,11 @@
 ./tests/models/test_dagbag.py
 ./tests/models/test_dagrun.py
 ./tests/models/test_pool.py
-./tests/models/test_skipmixin.py
 ./tests/models/test_taskinstance.py
 ./tests/operators/test_bash_operator.py
 ./tests/operators/test_cassandra_to_gcs.py
 ./tests/operators/test_gcs_to_s3.py
 ./tests/operators/test_hive_operator.py
-./tests/operators/test_hive_to_druid.py
 ./tests/operators/test_mssql_to_hive.py
 ./tests/operators/test_operators.py
 ./tests/operators/test_papermill_operator.py
@@ -342,8 +326,7 @@
 ./tests/operators/test_s3_to_redshift_operator.py
 ./tests/operators/test_virtualenv_operator.py
 ./tests/plugins/test_plugin.py
-./tests/providers/amazon/aws/hooks/test_redshift.py
-./tests/providers/amazon/aws/sensors/test_redshift.py
+./tests/providers/apache/cassandra/hooks/test_cassandra_hook.py
 ./tests/security/test_kerberos.py
 ./tests/sensors/test_base_sensor.py
 ./tests/sensors/test_external_task_sensor.py
@@ -358,22 +341,17 @@
 ./tests/test_configuration.py
 ./tests/test_local_settings.py
 ./tests/test_logging_config.py
-./tests/test_stats.py
 ./tests/test_utils/reset_warning_registry.py
 ./tests/ti_deps/deps/fake_models.py
-./tests/ti_deps/deps/test_dagrun_exists_dep.py
 ./tests/ti_deps/deps/test_ready_to_reschedule_dep.py
-./tests/ti_deps/deps/test_runnable_exec_date_dep.py
 ./tests/utils/log/test_es_task_handler.py
 ./tests/utils/log/test_file_processor_handler.py
 ./tests/utils/log/test_s3_task_handler.py
 ./tests/utils/test_cli_util.py
 ./tests/utils/test_compression.py
 ./tests/utils/test_dag_processing.py
-./tests/utils/test_dates.py
 ./tests/utils/test_db.py
 ./tests/utils/test_decorators.py
 ./tests/utils/test_helpers.py
 ./tests/utils/test_log_handlers.py
 ./tests/utils/test_tests.py
-./tests/utils/test_timezone.py
